requirements.txt
python-dotenv>=1.0.0
google-adk-agents
lite-llm
